Added max ladder fuel input value, to prevent high ladder fuels from causing elevated severity even in mild fire weather. Added ladder fuel output layer for calibration

+++ src/FireEvent.cs
@@ -288,7 +288,7 @@ namespace Landis.Extension.Scrapple
             }
 
             // LADDER FUELS ************************
-            double ladderFuelBiomass = SiteVars.LadderFuels[site];
+            int ladderFuelBiomass = SiteVars.LadderFuels[site];
 
             
             // End LADDER FUELS ************************
